Replace QTextEdit with QPlainTextEdit
Formatting is not really needed at the current time. This changes makes
the API more clean because we don't have to store the html to keep the
formatting as before. The plaintext now contains the newlines we need.

Furthermore adds a version field when exporting. This makes it easier
to deserialize different versions of reports.

fixes #12